Refactor get(_unchecked) to always return a PixelKind
This makes the interfaces more consistent, and makes it such that there
is exactly one way to get the data. By doing the cast after determining
the format kind, we can guarantee that we are referencing into the
correct type. The code is still unsafe (the compiler cannot guarantee
that the underlying memory is still valid since we're getting a
reference to a C pointer across the FFI) but at least we won't be doing
any intermediate copies!