Refactor sensor class
There's some big changes here, and this breaks a lot of the existing
library, which adds more TODOs. Anyways, the main changes:

- Sensor is no longer a template type. There is nothing about the type
or its interfaces that require you to care about the "Kind" subtype. In
particular, get_option and info both allow you to get any information
about a particular sensor object.
- As corollary to above, the ExtendedSensor type is no longer needed.
- Sensors _can_ have multiple extensions, which you can now query using
`extensions(&self)`.
- Stream profiles no longer returns a stream profile list type, but
rather just a vector of stream profiles. It's a bit of extra work at
every call site, but the ergonomics of a native Rust type justify it.