Add lifetime to StreamProfile
Needed because otherwise users might decide to hold onto this type and
that would be a bad idea. Lifetimes for this type are managed by the
device. Even if you get the frame stream profile, you _don't_ drop it,
and it isn't necessarily going to be valid. An example where a stream
profile may be invalid:

1. Connect device
2. Stream data
3. Get frame
4. Disconnect device (unplug from your computer)
5. Ask for stream profile from frame

Since the device is disconnected (and ostensibly de-allocated) we can no
longer rely on the stream profile from the frame. Since the rust
implementation caches parts of it (notably format) we can continue to
use that even if the pointer is invalid. This is useful to work around
errors like in the above example.

For this reason, we add a lifetime to the profile, which more or less
denotes that it can't outlive it's parent object. This does prohibit one
from calling get/set extrinsics or get intrinsics, so perhaps the
getters for these should be cached as well?