Add an API for getting an owned, non-null frame pointer
This is necessary because some parts of the low-level C API want to take
ownership of your frame pointer, but you need a safe way to get that
without keeping the frame object itself alive. By having this API take
ownership of `self`, we're able to provide a way for you to "consume"
your frame object, and then return just the pointer. This sets an
internal flag on these types so that they don't get dropped and we don't
call `sys::rs2_release_frame` on the frame pointer. Instead, the frame
type is released but the frame is still valid.