Add more complete set of Rs2Option APIs on `Sensor`
While documenting I realized that the API didn't provide setters, nor
did the API provide any means for the user to get the range for an
option.

Predicates for `supports_option` and `is_option_read_only` were added so
that users can directly ask the API if getting / setting an option is
possible for the current sensor.