Remove old error checking code
The macro is sufficient for 90% of cases. I also question some of the
utility of the error type's implementation. For example:

* Implementing send and sync for an error type seems misguided. Errors
typically should be handled local to a thread, or signaled via channel.
In cases where that's not possible, `Arc::new` exists, and there isn't a
strong case for mutating an error.

* Compared to the macro which directly checks the error in place, it
seemed odd to me to try and RAII the error pointer and then signal some
kind of panic if you forget to check the result. Careful coding, etc.,
but it provides marginal benefit IMO for a type that is un-ergonomic.

* Having a single error type for every possible error in the crate is
also a problem. While rs2_error is more or less a thin C wrapper around
a C++ exception, the crate itself can at least attempt to be more
descriptive and separate different types of construction errors from
null pointer de-refs.

The check_rs2_error macro enables the same effective check, can easily
interoperate with e.g. the anyhow crate, and doesn't categorize all
errors under the same umbrella.