Refactor InactivePipeline
This simplifies the API greatly. Some food for thought:

- Active and inactive pipelines are now two separate types (rather than
instances of a meta-type). At their core they are the same type, but the
reason for distinguishing between two types is to provide API separation
between APIs that can be called before the pipeline is started
(streaming active) vs. when the pipeline is stopped (no streaming).

- InactivePipeline::new was removed. I don't see a good reason to have
the separate constructor. The underlying C++ API takes in a shared
pointer to a context (which is why we represent the context as a &'a ref
in Rust). Hiding this seems fishy since the user should be aware of the
context (and its lifetime) in their code. The one constructor that takes
in a Context object makes the most sense. The only question is: when
calling get_raw() on either context or config -- should we consume the
object (i.e. pass self or &self)? For Config I suspect the answer is
consume self, but for Context consume &self.

- Enforcing a specific type of async where we spawn a whole thread just
to "start" the pipeline doesn't seem like the optimal solution. If you
wanted to integrate with tokio or async-std, there are APIs for dealing
with this and can save you the cost of spawning your own thread.

- Breaking the pipeline into raw parts isn't necessary. The "from raw
parts" is really a TryFrom, so that has been converted. However,
breaking a pipeline into the underlying raw pointer seems foolish. It's
not an API necessity and this really isn't something users should reach
for. We should instead strive for complete pipeline API coverage.

Co-authored-by: Brandon Minor <brandon@tangramvision.com>