Remove ToOption trait from crate
This is a trait that only really matters for two types:

- Sensor
- ProcessingBlock

In both cases, we never want to export this trait. You should export a
trait if:

1. Users will be re-implementing this trait onto their own types
2. This trait abstracts over types that would be used in the same
   context.

Sensor and ProcessingBlock are unlikely to be leveraging any kind of
sub-type polymorphism / sub-type relationship, so the trait isn't really
worth much. Obviously these methods need to be re-introduced to
ProcessingBlock when that is re-enabled in the crate, but for now this
trait is unnecessary.