Remove futures and tokio as (dev) dependencies
Going forward I don't think it's particularly helpful to bundle these in
the library ourself. Main justification is that users may not like the
async library we choose, and it's easy enough for them to wrap this in
their own wrappers.

Moreover, the way that async was done prior to this removal (as well as
before pipeline was refactored) was to spawn a thread, and then await
the detached thread on a one-use channel. This... isn't always ideal.
More in depth:

- InactivePipeline::start_async() is not an inherently asynchronous
function. There's no IO, and assuming you check your device hub, you
shouldn't be blocking on it (at least, it should return as soon as the
config is resolved).

- ActivePipeline::wait_async() is even less ideal: every time you wait
for the next frame set, you spawn a new thread. This is pretty
heavyweight and expensive, especially if you're waiting on a RealSense
IMU that might be outputting data at (much higher than 30 FPS) frame
rate.

Since this is really meant to be a wrapper around librealsense2, we
should probably not expect async included by default, since there
doesn't (yet) exist an API which handles frame grabbing in an async way
(using epoll, io_uring, etc). Instead, if you're looking to wrap things
in tokio or async-std, you're better off using their "blocking"
executors. Generally, those will be better at managing the thread
overhead than we can by hiding a detached thread inside async methods.

If you want an example of how to do this with async-std, implement the
"unstable" feature and do roughly the following:

```
use async_std::task;

let pipeline = // ... set this up as usual

task::spawn_blocking(move {
    pipeline.wait(None)?
}).await
```

This way, you put blocking tasks on specially scheduled threads, and
don't hang the rest of your system. Note: you could just as easily spawn
a single thread yourself and have it continuously run the above in a
loop, and have that push to an async channel of some kind. This is
probably what you ought to do if you want more than a single frame at a
time and you want to ensure this loop doesn't block.

In short: threading is hard, but trying to provide "nice" async
implementations on top of an API that hard blocks is sub-optimal at
best, and actively misleading users of the library at worst (since they
probably will assume we thought of all these trade-offs and can just use
it as it's packaged).

Instead, we just remove any async capabilities and hope to instruct /
document this behaviour through live examples when possible.