Fix dumb clippy error
This one is pretty dumb. The type is copy & clone, and if we don't call
clone, it complains we're returning a reference. If we do, it complains
we are cloning a copy type. So I guess the extra code in a weird style
justifies it?

It's not the worst workaround and it occurs in exactly one place so I'll
permit it, but this is why people disable linters and think they're
braindead (even if linters often more useful than hurtful).