Add type-specificity via FrameCategory trait
librealsense2 returns all observations (computer vision or processed) as
pointers to opaque structs. Because of this, types are flexible.
However, this also means that understanding the data format, layout,
etc. requires a lot of work.

One of the principles of the high-level realsense-rust library is to try
and keep things strongly typed while maintaining that flexibility. This
is done to try and increase expressiveness while handling memory safety
under the hood.

There are three components to an RS2 frame, which we use to define a
frame's "type" in Rust:

1. The format (see Rs2Format) of the data
2. The interfaces (see Rs2Extension) that the frame can be "extended" to
3. The stream kind, which defines a "category" across frames

The former is more or less handled by our `PixelKind` type (and in some
cases, you only have one or two options for data format for a given
frame e.g. MotionXyz32f).

The latter two components are less dynamic. These are necessary
distinctions to make when trying to filter specific frames from e.g. a
CompositeFrame (which is returned from the pipeline).

These components have been bundled together into a FrameCategory trait,
which should uniquely define the "category" of frame that a type
implement. This is done so that asking for e.g. frames that satisfy the
"video frame" extension doesn't also return depth frames.

We now have several more frame types than before, with more specificity:

- ColorFrame
- InfraredFrame
- PoseFrame
- DepthFrame
- DisparityFrame
- PointsFrame
- ConfidenceFrame
- FisheyeFrame
- AccelFrame
- GyroFrame

This distinguishes between all the possible frame combinations that can
be returned by librealsense2.

Users should probably not implement the FrameCategory trait themselves
unless they're doing something sneaky.