Remove rstuil_delegate from realsense-sys
This is explicitly removing functionality, but explicitly removing
functionality that isn't very useful. Here's the gist:

1. rsutil.h isn't particularly useful. This crate has never provided
high level bindings for the functions therein. These functions are also
fairly short and don't provide any unique functionality that is
necessary to use the crate.

2. rsutil.h defines everything as static functions, with inline
definitions. This is why there's the whole need for the rsutil_delegate
scaffolding. The dependency on the CC crate is removed by removing this
functionality.

3. The rsutil functionality does break in librealsense2 v2.42.0 because
someone used a functional cast expression in one of the static methods.
This would work in a C++ compiler but not a C compiler.

In summary, the functionality provided by this is extremely limited and
it's a lot of random code to maintain. We'd be better served writing our
own functions in pure Rust (type safety) and tagging them with the
\#[inline] attribute. It simplifies the build system and provides better
guarantees for the realsense-rust crate & high-level bindings.

Please complain if this functionality is needed! I'm happy to re-write
these as pure Rust functions if there's a desire for them.