Move size calculation to separate statement
This is done to sneak this past a clippy lint. In this particular
instance, I know what I'm doing. The lint expects this when you do

usize * sizeof(T)

But I'm doing

size / sizeof(T)

I disagree with clippy here, this is the correct thing to do since
librealsense2 gives us the size in bytes, but we want the number of
f32s.