Cast all enum types to u32
Windows bindings for RealSense introducs a small change from the Linux version:
enums go from being u32 to i32. This is a _big_ change for this library, since
we had assumed up to this point that everything was u32 or eventually converted
to u32.

This is made more complicated by the fact that enum declarations must be made
with a constant object. This means that I could not employ any `try_from` or
`u32::from` logic to make the unification between platforms.

I did consider using something along these lines:

```
[cfg_attr(target_os = "windows", repr(i32))]
[cfg_attr(not(target_os = "windows"), repr(u32))]
```

...and just having enums be i32 on Windows. However, I realized the downstream
effects of this would be rough; everything using an enum would have to do
something along these lines, as the type wouldn't be constant.

I opted to do what I did (the excessive use of `as u32`) because it met the req
of being constant, while also maintaining the constant cast of the enums between
platforms. There are still some funky pieces of code that result, mainly the
`as_u32().unwrap().try_from().unwrap()`, which is trying _really hard_ to make
sure the casting is good to go.