Correct simplified build script for cross-platform compilation
This work was also done on the original repository started by @jerry73204,
contributed by user @de-vri-es. It was an attempt at simplifying the build
script and potentially bypassing pkg-config when not needed. However, the
rsutils helper compliation step was broken, as was the logic for finding the
header when building outside of realsense-sys. These have been corrected, and
this build script is working pretty well for Windows.