Add DLL file to executable path if available (Windows)
Rulst uses .lib files to create a binary, but .dll files on runtime. This means
that linking the .lib files alone isn't enough, and since the .dll needs to be
present, moving the .dll wholesale is the easiest way.

After struggling about it for a bit, I finally decided to add the DLL directory
as an extra argument via pkg-config. This allows for flexibility in search,
while also maintaining the use of pkg-config as the main linking tool.

The copy logic here is admittedly not graceful, but it works consistently between
building realsense-sys and realsense-rust.