Add test on l500 and d400 series for supported but ignored options
Some of the options available to set on a sensor are actually not supported at
all, or are overwritten without the user's knowledge. This is only the case for
GlobalTimeEnabled on the L500 series, but tests have been written to easily
iterate over any setting that might be in question.

This commit also splits the device-specific tests into their own files, rather
than stacking them all into one.