Move the DESIGN readme into a docs/ rust directory for rustdoc consumption
I'm of two minds with this commit. On the one hand, it utilizes rustdocs in a
way that I find suitable for the library. It's well-marked as a doc module, and
there's no extra code that would interfere if it was imported for some reason.

On the other hand, I've never seen it done. It's true that this could be a
README, but rustdoc does such a good job with internal linking in the codebase
that I'd rather use the tools that I'm given.