Convert all kind enums from u32 to i32
It appears as if bindgen chooses i32 always for Windows, but varies for Mac and
Linux. Relevant links:

- https://users.rust-lang.org/t/bindgen-and-enum-types/34370
- https://github.com/rust-lang/rust-bindgen/issues/1361

However, this behavior isn't in line with the C standard, which states:

"The identifiers in an enumerator list are declared as constants that have type
int and may appear wherever such are permitted."

To remedy this for all platforms, enums are now default type i32.