Remove owned / borrowed Context object from pipelines
The Context object in this scenario was held to try and designate some
kind of lifetime relationship between pipelines and the Context itself.
Unfortunately, this leads to some very wonky semantics with `Send`ing
pipelines across threads.

This is especially bad if you're encountering an async executor, which
may await on one thread and return on another. `Arc<Context>` won't be
`Send` because `Context` isn't `Send`. And holding a reference is just
as bad because then the Pipeline and Context would have to be passed
together (and owned). Since multiple pipelines can exist per context,
this is just frustrating.

At a lower level the context objects used in pipelines are held as
`std::shared_ptr<T>`. It is unclear if this is maintained in any sane
way with regards to the whole program, since getting a new context seems
to effectively call `std::shared_ptr<Context>::get()`, which seems
incredibly unsafe.

For now we will instead relax the restriction on pipelines holding
Context types, and revisit in the future if there is a compelling reason
to do so.

One alternative may be that holding onto `Arc<Mutex<Context>>` would
make sense, but then we have to commit to a specific `Mutex` type,
unless we make our pipeline types even more complex by templatizing
them. I'd rather not force this choice onto the user since in most cases
they will probably just get confused for very little benefit.