Remove the lifetime from stream profiles
The rationale for this is that this propagates explicit lifetimes for
like half the API, in particular the ActivePipeline, which makes working
with the data wonky. The initial motivation for this was to prevent
users from holding onto a stream profile beyond the lifetime of the
owning object (profile list or frame).

For frames, we can control this by never returning an owned profile, and
instead just providing a reference. You probably shouldn't be using this
profile much anyways outside of getting the data format for image
frames.

For pipeline profiles, this is admittedly still a little wonky, but not
too bad. We return a cached vector rather than an exact copy of every
stream profile, even though because we generate the vector from a
stream profile list we get a full set of owned sensor profiles now.

If you query the stream list from a sensor type however, you can get a
fully owned vector of all stream profiles.

The final moral here is that we don't need the artificial phantom data +
explicit lifetime that was encoded into stream profiles before, instead
relying on actual ownership. It does mean whenever a stream profile is
constructed from a rs2_stream_profile_list (via `try_create`) we are
doing a clone of the data via the rs2 API. However, this just makes the
ownership semantics much easier to understand and requires less explicit
lifetime management, so I see it as a net-win across the API.

Side note: it seems strange to me that there is three ways to get this
data. The RS2 API continues to confound and elude reason.